# Decouples Brave from zipkin2.Endpoint