--- a/brave/src/main/java/brave/internal/recorder/MutableSpan.java
+++ b/brave/src/main/java/brave/internal/recorder/MutableSpan.java
@@ -2,10 +2,12 @@ package brave.internal.recorder;
 
 import brave.Span.Kind;
 import brave.Tracer;
+import brave.internal.IpLiteral;
+import brave.internal.Nullable;
 import brave.propagation.TraceContext;
 import java.util.ArrayList;
 import java.util.List;
-import zipkin2.Endpoint;
+import java.util.Locale;
 import zipkin2.Span;
 
 /**
@@ -22,8 +24,9 @@ public final class MutableSpan {
   Kind kind;
   boolean shared;
   long startTimestamp, finishTimestamp;
-  String name;
-  Endpoint localEndpoint, remoteEndpoint;
+  String name, remoteServiceName, remoteIp;
+  int remotePort;
+
   /**
    * To reduce the amount of allocation, collocate annotations with tags in a pair-indexed list.
    * This will be (startTimestamp, value) for annotations and (key, value) for tags.
@@ -45,6 +48,50 @@ public final class MutableSpan {
   public void kind(Kind kind) {
     if (kind == null) throw new NullPointerException("kind == null");
     this.kind = kind;
+  }
+
+  /** @see brave.Span#remoteServiceName(String) */
+  @Nullable public String remoteServiceName() {
+    return remoteServiceName;
+  }
+
+  /** @see brave.Span#remoteServiceName(String) */
+  public void remoteServiceName(String remoteServiceName) {
+    if (remoteServiceName == null || remoteServiceName.isEmpty()) {
+      throw new NullPointerException("remoteServiceName is empty");
+    }
+    this.remoteServiceName = remoteServiceName.toLowerCase(Locale.ROOT);
+  }
+
+  /**
+   * The text representation of the primary IPv4 or IPv6 address associated with the remote side of
+   * this connection. Ex. 192.168.99.100 null if unknown.
+   *
+   * @see brave.Span#remoteIpAndPort(String, int)
+   */
+  @Nullable public String remoteIp() {
+    return remoteIp;
+  }
+
+  /**
+   * Port of the remote IP's socket or 0, if not known.
+   *
+   * @see java.net.InetSocketAddress#getPort()
+   * @see brave.Span#remoteIpAndPort(String, int)
+   */
+  public int remotePort() {
+    return remotePort;
+  }
+
+  /** @see brave.Span#remoteIpAndPort(String, int) */
+  public boolean remoteIpAndPort(@Nullable String remoteIp, int remotePort) {
+    if (remoteIp == null) return false;
+    this.remoteIp = IpLiteral.ipOrNull(remoteIp);
+    if (this.remoteIp == null) return false;
+    if (remotePort > 0xffff) throw new IllegalArgumentException("invalid port " + remotePort);
+    if (remotePort < 0) remotePort = 0;
+    this.remotePort = remotePort;
+    return true;
   }
 
   /** @see brave.Span#annotate(String) */
@@ -70,18 +117,6 @@ public final class MutableSpan {
     pairs.add(value);
   }
 
-  /** @see brave.Tracing.Builder#endpoint */
-  public void localEndpoint(Endpoint localEndpoint) {
-    if (localEndpoint == null) throw new NullPointerException("localEndpoint == null");
-    this.localEndpoint = localEndpoint;
-  }
-
-  /** @see brave.Span#remoteEndpoint(Endpoint) */
-  public void remoteEndpoint(Endpoint remoteEndpoint) {
-    if (remoteEndpoint == null) throw new NullPointerException("remoteEndpoint == null");
-    this.remoteEndpoint = remoteEndpoint;
-  }
-
   /**
    * Indicates we are contributing to a span started by another tracer (ex on a different host).
    * Defaults to false.
@@ -101,16 +136,22 @@ public final class MutableSpan {
   // Since this is not exposed, this class could be refactored later as needed to act in a pool
   // to reduce GC churn. This would involve calling span.clear and resetting the fields below.
   public void writeTo(zipkin2.Span.Builder result) {
-    result.localEndpoint(localEndpoint);
-    result.remoteEndpoint(remoteEndpoint);
     result.name(name);
+    if (kind != null && kind.ordinal() < Span.Kind.values().length) { // defend against version skew
+      result.kind(zipkin2.Span.Kind.values()[kind.ordinal()]);
+    }
     result.timestamp(startTimestamp);
     if (startTimestamp != 0 && finishTimestamp != 0L) {
       result.duration(Math.max(finishTimestamp - startTimestamp, 1));
     }
-    if (kind != null && kind.ordinal() < Span.Kind.values().length) { // defend against version skew
-      result.kind(zipkin2.Span.Kind.values()[kind.ordinal()]);
+    if (remoteIp != null || remoteServiceName != null) {
+      result.remoteEndpoint(zipkin2.Endpoint.newBuilder()
+          .serviceName(remoteServiceName)
+          .ip(remoteIp)
+          .port(remotePort)
+          .build());
     }
+
     for (int i = 0, length = pairs.size(); i < length; i += 2) {
       Object first = pairs.get(i);
       String second = pairs.get(i + 1).toString();
@@ -123,6 +164,6 @@ public final class MutableSpan {
     if (shared) result.shared(true);
   }
 
-  MutableSpan(){ // intentionally hidden
+  MutableSpan() { // intentionally hidden
   }
 }