This adds methods that decouple Brave apis from zipkin2.Endpoint:
* Tracing.Builder.localIp, localPort for global config
* Span.remoteServiceName, remoteIpAndPort for remote endpoint

Fixes #745