@adriancole 2018-08-04
I owe some more tests and benchmarks, but currently happy with the shape of this.