@adriancole 2018-08-05
sometimes we use the word maybe.. ex `maybeRemoteIpAndPort` which returns true if it is, could be more sensible than `parseRemoteIpAndPort`.. especially as we usually use the word parse when scraping types like HttpRequest, and in this case it is just a string and int. 

@jcchavezs @devinsba @marcingrzejszczak @llinder @devinsba @shakuzen @shimamoto any preferences? I agree this method is the odd-ball, definitely need to pick something sensible.

Here are some choices with real code. I didn't add "AndSet" as there are too many ands. Plus we don't use the word "set" in any of our code, but I could be convinced.:

Parse (current)
```java
span.remoteServiceName("backend");
span.parseRemoteIpAndPort("172.3.4.1", 8108);

  /**
   * This sets the client IP:port to the {@linkplain HttpServletRequest#getRemoteAddr() remote
   * address} if the {@link HttpServerAdapter#parseClientIpAndPort default parsing} fails.
   */
  @Override public boolean parseClientIpAndPort(HttpServletRequest req, Span span) {
    if (super.parseClientIpAndPort(req, span)) return true;
    return span.parseRemoteIpAndPort(req.getRemoteAddr(), req.getRemotePort());
  }
```

Maybe (different word than parse)
```java
span.remoteServiceName("backend");
span.maybeRemoteIpAndPort("172.3.4.1", 8108);
  /**
   * This sets the client IP:port to the {@linkplain HttpServletRequest#getRemoteAddr() remote
   * address} if the {@link HttpServerAdapter#parseClientIpAndPort default parsing} fails.
   */
  @Override public boolean parseClientIpAndPort(HttpServletRequest req, Span span) {
    if (super.parseClientIpAndPort(req, span)) return true;
    return span.maybeRemoteIpAndPort(req.getRemoteAddr(), req.getRemotePort());
  }
```

OrFalse? (allows you to read the methods lexicographically and have a suffix telling you it is different). Might be a different suffix choice
```java
span.remoteServiceName("backend");
span.remoteIpAndPortOrFalse("172.3.4.1", 8108);

  /**
   * This sets the client IP:port to the {@linkplain HttpServletRequest#getRemoteAddr() remote
   * address} if the {@link HttpServerAdapter#clientIpAndPortOrFalse default parsing} fails.
   */
  @Override public boolean clientIpAndPortOrFalse(HttpServletRequest req, Span span) {
    if (super.clientIpAndPortOrFalse(req, span)) return true;
    return span.remoteIpAndPortOrFalse(req.getRemoteAddr(), req.getRemotePort());
  }
```