@adriancole 2018-08-05
PS so far I am still fine with the parse prefix as it is used everywhere else, although only zipkin2.Endpoint returns boolean when doing so.