@adriancole 2018-08-05
So far, I think this is a performance win outside statistical error, even if only a few microseconds

Before
```
Benchmark                                            Mode  Cnt    Score   Error  Units
EndToEndBenchmarks.server_get                        avgt   15  142.404 ± 4.273  us/op
EndToEndBenchmarks.traced128Server_get               avgt   15  198.336 ± 2.636  us/op
EndToEndBenchmarks.traced128Server_get_resumeTrace   avgt   15  202.861 ± 4.952  us/op
EndToEndBenchmarks.tracedExtraServer_get             avgt   15  210.209 ± 5.718  us/op
EndToEndBenchmarks.tracedExtraServer_get_request_id  avgt   15  213.879 ± 3.727  us/op
EndToEndBenchmarks.tracedServer_get                  avgt   15  199.546 ± 3.893  us/op
EndToEndBenchmarks.tracedServer_get_resumeTrace      avgt   15  207.808 ± 5.562  us/op
EndToEndBenchmarks.unsampledServer_get               avgt   15  157.828 ± 3.170  us/op
```

After
```
Benchmark                                            Mode  Cnt    Score    Error  Units
EndToEndBenchmarks.server_get                        avgt   15  142.165 ±  3.012  us/op
EndToEndBenchmarks.traced128Server_get               avgt   15  193.668 ±  3.397  us/op
EndToEndBenchmarks.traced128Server_get_resumeTrace   avgt   15  195.480 ±  2.659  us/op
EndToEndBenchmarks.tracedExtraServer_get             avgt   15  206.897 ± 13.223  us/op
EndToEndBenchmarks.tracedExtraServer_get_request_id  avgt   15  210.645 ±  5.961  us/op
EndToEndBenchmarks.tracedServer_get                  avgt   15  191.812 ±  1.660  us/op
EndToEndBenchmarks.tracedServer_get_resumeTrace      avgt   15  197.930 ±  6.315  us/op
EndToEndBenchmarks.unsampledServer_get               avgt   15  152.701 ±  2.938  us/op
```