@adriancole 2018-08-05
renamed Span.parseRemoteIpAndPort to remoteIpAndPort to avoid the distraction. lemme know if works for you @jcchavezs!