@adriancole 2018-08-06
revised description